Fixed a compiler warning with gcc 4.3. Fixes LM-127
Patch from Frederic Peters.

+++ loudmouth/lm-ssl-gnutls.c
@@ -212,7 +212,7 @@ _lm_ssl_begin (LmSSL *ssl, gint fd, const gchar *server, GError **error)
 				ssl->gnutls_xcred);
 
 	gnutls_transport_set_ptr (ssl->gnutls_session,
-				  (gnutls_transport_ptr_t) fd);
+				  (gnutls_transport_ptr_t)(glong) fd);
 
 	ret = gnutls_handshake (ssl->gnutls_session);
